API Reference

NeuralAttentionlib.AbstractAttenMask

NeuralAttentionlib.AbstractAttenMaskOp

NeuralAttentionlib.apply_mask

NeuralAttentionlib.apply_mask

NeuralAttentionlib.collapsed_size

NeuralAttentionlib.naive_qkv_attention

NeuralAttentionlib.AbstractAttenMask — Type

Wrapper type for mask data, can be viewed as AbstractArray{Bool}

NeuralAttentionlib.AbstractAttenMaskOp — Type

Trait-like type for holding operation related argument, defined how the mask should be apply to input array

NeuralAttentionlib.apply_mask — Method

Equivalent to op.apply(score, op.scale .* (op.flip ? .! mask : mask)).

For example: apply_generic_mask(GenericAttenMaskOp(.+, static(true), -1e9), mask, score) == @. score + (!mask * -1e9).

NeuralAttentionlib.apply_mask — Method

Directly broadcast multiply mask to attention score.

NeuralAttentionlib.collapsed_size — Method

collapsed_size(x, xi, xj)

Collapse the dimensionality of x into 3 according to xi and xj.

(X1, X2, ..., Xi-1, Xi, Xi+1, ..., Xj-1, Xj, ..., Xn)
 |_____dim1______|  |_______dim2______|  |___dim3__|

This is equivalent to size(reshape(x, prod(size(x)[1:(xi-1)]), prod(size(x)[xi:(xj-1)]), prod(size(x)[xj:end]))).

#Example

julia> x = randn(7,6,5,4,3,2);

julia> collapsed_size(x, 3,5)
(42, 20, 6)

NeuralAttentionlib.naive_qkv_attention — Function

equivalent to generic_qkv_attention(weighted_sum_mixing, normalized_score(NNlib.softmax) $ masked_score(mask) $ scaled_dot_product_score, q, k, v)