API Reference

Base.:!

Base.:&

Base.:|

NeuralAttentionlib.apply_mask

NeuralAttentionlib.apply_mask

NeuralAttentionlib.collapsed_size

NeuralAttentionlib.noncollapsed_size

NeuralAttentionlib.AbstractArrayMask

NeuralAttentionlib.AbstractAttenMask

NeuralAttentionlib.AbstractAttenMaskOp

NeuralAttentionlib.AbstractDatalessMask

NeuralAttentionlib.BandPartMask

NeuralAttentionlib.BatchedMask

NeuralAttentionlib.BiLengthMask

NeuralAttentionlib.CausalMask

NeuralAttentionlib.CollapsedDimArray

NeuralAttentionlib.GenericMask

NeuralAttentionlib.LocalMask

NeuralAttentionlib.RandomMask

NeuralAttentionlib.RepeatMask

NeuralAttentionlib.SymLengthMask

Mask

NeuralAttentionlib.AbstractAttenMask — Type

AbstractAttenMask

Abstract type for mask data, can be viewed as AbstractArray{Bool}

NeuralAttentionlib.AbstractAttenMaskOp — Type

AbstractAttenMaskOp

Trait-like abstract type for holding operation related argument, defined how the mask should be apply to input array

NeuralAttentionlib.apply_mask — Method

apply_mask(op::GenericAttenMaskOp, mask::AbstractAttenMask, score)

Equivalent to op.apply(score, op.scale .* (op.flip ? .! mask : mask)).

Example

julia> x = randn(10, 10);

julia> m = CausalMask()
CausalMask()

julia> apply_mask(GenericAttenMaskOp(.+, true, -1e9), m, x) ==  @. x + (!m * -1e9)
true

NeuralAttentionlib.apply_mask — Method

apply_mask(op::NaiveAttenMaskOp, mask::AbstractAttenMask, score)

Directly broadcast multiply mask to attention score, i.e. score .* mask.

NeuralAttentionlib.AbstractArrayMask — Type

AbstractArrayMask <: AbstractAttenMask

Abstract type for mask with array data

NeuralAttentionlib.AbstractDatalessMask — Type

AbstractDatalessMask <: AbstractAttenMask

Abstract type for mask without array data.

NeuralAttentionlib.BandPartMask — Type

BandPartMask(l::Int, u::Int) <: AbstractDatalessMask

Attention mask that only allow band_part values to pass.

NeuralAttentionlib.BatchedMask — Type

BatchedMask(mask::AbstractArrayMask, batch_dim::Int) <: AbstractWrapperMask

Attention mask wrapper over array mask for applying the same mask within the same batch. Use batch_dim to specify from which dim should be treated as batch dims.

Example

julia> m = SymLengthMask([2,3])
SymLengthMask{1, Vector{Int32}}(Int32[2, 3])

julia> trues(3,3, 2) .* m
3×3×2 BitArray{3}:
[:, :, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 2] =
 1  1  1
 1  1  1
 1  1  1

julia> trues(3,3, 2, 2) .* m
ERROR: [...]

julia> trues(3,3, 2, 2) .* BatchedMask(m, 4) # 4-th dim is batch dim
3×3×2×2 BitArray{4}:
[:, :, 1, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 2, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 1, 2] =
 1  1  1
 1  1  1
 1  1  1

[:, :, 2, 2] =
 1  1  1
 1  1  1
 1  1  1

# can also use negative value to count from the last
julia> trues(3,3, 2, 2) .* BatchedMask(m, 4) == trues(3,3, 2, 2) .* BatchedMask(m, -1)
true

NeuralAttentionlib.BiLengthMask — Type

BiLengthMask(q_len::A, k_len::A) where {A <: AbstractArray{Int, N}} <: AbstractArrayMask

Attention mask specified by two arrays of integer that indicate the length dimension size.

Example

julia> bm = BiLengthMask([2,3], [3, 5])
BiLengthMask{1, Vector{Int32}}(Int32[2, 3], Int32[3, 5])

julia> trues(5,5, 2) .* bm
5×5×2 BitArray{3}:
[:, :, 1] =
 1  1  0  0  0
 1  1  0  0  0
 1  1  0  0  0
 0  0  0  0  0
 0  0  0  0  0

[:, :, 2] =
 1  1  1  0  0
 1  1  1  0  0
 1  1  1  0  0
 1  1  1  0  0
 1  1  1  0  0

See also: SymLengthMask, BatchedMask, RepeatMask

NeuralAttentionlib.CausalMask — Type

CausalMask() <: AbstractDatalessMask

Attention mask that block the future values.

Similar to applying LinearAlgebra.triu! on the score matrix

NeuralAttentionlib.GenericMask — Type

GenericMask <: AbstractArrayMask

Generic attention mask. Just a wrapper over AbstractArray{Bool} for dispatch.

NeuralAttentionlib.LocalMask — Type

LocalMask(width::Int) <: AbstractDatalessMask

Attention mask that only allow local (diagonal like) values to pass.

width should be ≥ 0 and A .* LocalMask(1) is similar to Diagonal(A)

NeuralAttentionlib.RandomMask — Type

RandomMask(p::Float64) <: AbstractDatalessMask

Attention mask that block value randomly.

p specify the percentage of value to block. e.g. A .* RandomMask(0) is equivalent to identity(A) and A .* RandomMask(1) is equivalent to zero(A).

NeuralAttentionlib.RepeatMask — Type

RepeatMask(mask::AbstractAttenMask, num::Int) <: AbstractWrapperMask

Attention mask wrapper over array mask for doing inner repeat on the last dimension.

Example

julia> m = SymLengthMask([2,3])
SymLengthMask{1, Vector{Int32}}(Int32[2, 3])

julia> trues(3,3, 2) .* m
3×3×2 BitArray{3}:
[:, :, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 2] =
 1  1  1
 1  1  1
 1  1  1

julia> trues(3,3, 4) .* m
[...] # wrong result due to out of bound access

julia> trues(3,3, 4) .* RepeatMask(m, 2)
3×3×4 BitArray{3}:
[:, :, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 2] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 3] =
 1  1  1
 1  1  1
 1  1  1

[:, :, 4] =
 1  1  1
 1  1  1
 1  1  1

NeuralAttentionlib.SymLengthMask — Type

SymLengthMask(len::AbstractArray{Int, N}) <: AbstractArrayMask

Attention mask specified by an array of integer that indicate the length dimension size. assuming Query length and Key length are the same.

Example

julia> m = SymLengthMask([2,3])
SymLengthMask{1, Vector{Int32}}(Int32[2, 3])

julia> trues(3,3, 2) .* m
3×3×2 BitArray{3}:
[:, :, 1] =
 1  1  0
 1  1  0
 0  0  0

[:, :, 2] =
 1  1  1
 1  1  1
 1  1  1

See also: BiLengthMask, BatchedMask, RepeatMask

Base.:! — Method

!m::AbstractAttenMask

Boolean not of an attention mask

Base.:& — Method

m1::AbstractAttenMask & m2::AbstractAttenMask

logical and of two attention mask

Base.:| — Method

m1::AbstractAttenMask | m2::AbstractAttenMask

logical or of two attention mask

Matmul

NeuralAttentionlib.CollapsedDimArray — Type

CollapsedDimArray{T}(array, si::Integer=2, sj::Integer=3) <: AbstractArray{T, 3}

Similar to lazy reshape array with collapsed_size

NeuralAttentionlib.collapsed_size — Method

collapsed_size(x, xi, xj)::Dim{3}

Collapse the dimensionality of x into 3 according to xi and xj.

(X1, X2, ..., Xi-1, Xi, Xi+1, ..., Xj-1, Xj, ..., Xn)
 |_____dim1______|  |_______dim2______|  |___dim3__|

This is equivalent to size(reshape(x, prod(size(x)[1:(xi-1)]), prod(size(x)[xi:(xj-1)]), prod(size(x)[xj:end]))).

#Example

julia> x = randn(7,6,5,4,3,2);

julia> collapsed_size(x, 3, 5)
(42, 20, 6)

See also: noncollapsed_size

NeuralAttentionlib.noncollapsed_size — Method

noncollapsed_size(x, xi, xj, n)

Collapse the dimensionality of x into 3 according to xi and xj.

(X1, X2, ..., Xi-1, Xi, Xi+1, ..., Xj-1, Xj, ..., Xn)
 |_____dim1______|  |_______dim2______|  |___dim3__|

But take the size before collapse. e.g. noncollapsed_size(x, xi, xj, 2) will be (Xi, Xi+1, ..., Xj-1).

#Example

julia> x = randn(7,6,5,4,3,2);

julia> noncollapsed_size(x, 3, 5, 1)
(7, 6)

julia> noncollapsed_size(x, 3, 5, 2)
(5, 4)

julia> noncollapsed_size(x, 3, 5, 3)
(3, 2)

See also: collapsed_size